+++ cwl/src/main/scala/wdl4s/cwl/CwlToRun.scala
@@ -1,0 +1,18 @@
+package wdl4s.cwl
+
+import shapeless._
+
+object CwlToRun extends Poly1 {
+  implicit def commandLineTool =
+    at[CommandLineTool] {
+      clt =>
+       Coproduct[WorkflowStep.Run](clt)
+    }
+
+  implicit def workflow = at[Workflow] {
+    wf =>
+      Coproduct[WorkflowStep.Run](wf)
+  }
+}
+
+
