+++ cwl/src/main/scala/wdl4s/cwl/Decoder.scala
@@ -5,7 +5,7 @@ import ammonite.ops._
 import cats.effect.IO
 import shapeless.{Path => SPath, _}
 
-object Decoder {
+object CwlDecoder {
 
   type Parse[A] = EitherT[IO, NonEmptyList[String], A]
 
@@ -15,15 +15,30 @@ object Decoder {
 
   def lookupNamesIn(in: Cwl): Parse[Map[String, Cwl]] = ???
 
+  def traverseFileNames(in: String): StateT[IO, Cwl,Unit] = StateT{ s => IO P
+
+
+  def embedCwl(in: Workflow, map: Map[String, Cwl]): StateT[IO, Cwl, Unit] = {
+    //take thing that goes from
+    val filenames = in.steps.map(_.run.fold(
+
+
+      in.steps.run.select[String].toList.traverse(traverseFileNames)
+    }
+
   def modifyCwlWithEmbeddedCwl(in: Cwl, map: Map[String, Cwl]): Cwl =
     in.fold(AddEmbeddedCwl).apply(map)
 
+  /**
+   * Notice it gives you one instance of Cwl.  This has transformed all embedded files into scala object state
+   */
   def decodeAllCwl(filename: Path): Parse[Cwl] =
     for {
-      json <- salad(filename)
+      json <- salad(filename)//produces
       unmodifiedCwl <- parseJson(json)
-      modifyCwlWithEmbeddedCwl <- modifyCwlWithEmbeddedCwl(unmodifiedCwl)
-    } yield modifyCwlWithEmbeddedCwl
+      fileMap <- lookupNamesIn(unmodifiedCwl)
+      cwlWithEmbeddedCwl = modifyCwlWithEmbeddedCwl(unmodifiedCwl, fileMap)
+    } yield cwlWithEmbeddedCwl
 
 }
